index.jsをスッキリ
